split setup system
