+++ src/main.rs
@@ -1,6 +1,9 @@
 use bevy::prelude::*;
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
-use reve::constants::{Champion, GameInfo, Team, Textures, GAREN, MAP};
+use reve::{
+    constants::{Champion, GameInfo, Team, Textures, GAREN, MAP},
+    setup::SetupPlugin,
+};
 
 fn main() {
     App::new()
@@ -14,8 +17,8 @@ fn main() {
             ..Default::default()
         }))
         .add_plugin(WorldInspectorPlugin)
+        .add_plugin(SetupPlugin)
         .add_startup_system(setup)
-        .add_startup_system_to_stage(StartupStage::PostStartup, init_game)
         .run();
 }
 
@@ -55,33 +58,4 @@ fn setup(mut commands: Commands, asset: Res<AssetServer>) {
         garen: asset.load(GAREN),
     };
     commands.insert_resource(textures);
-}
-
-fn init_game(
-    mut commands: Commands,
-    textures: Res<Textures>,
-    mut cam_query: Query<&mut Transform, With<Camera2d>>,
-    game_info: Res<GameInfo>,
-) {
-    commands.spawn(SpriteBundle {
-        texture: textures.map.clone(),
-        transform: Transform {
-            translation: Vec3::new(0.0, 0.0, 1.0),
-            scale: Vec3::splat(1.), // TODO: Camera zoom
-            ..Default::default()
-        },
-        ..Default::default()
-    });
-
-    let mut cam = cam_query.get_single_mut().unwrap();
-    match game_info.team {
-        Team::Blue => {
-            cam.translation.x = -4800.0;
-            cam.translation.y = -4400.0;
-        }
-        Team::Red => {
-            cam.translation.x = 5000.0;
-            cam.translation.y = 3800.0;
-        }
-    }
 }
